// src/apps/what-to-do/models/activity-category.ts
export enum ActivityCategory {
    Outdoors = 'outdoors',
    Indoor = 'indoor',
    Sports = 'sports',
    Arts = 'arts',
    Music = 'music',
    Food = 'food',
    Travel = 'travel',
    Education = 'education',
    Health = 'health',
    Technology = 'technology',
    Entertainment = 'entertainment',
    Shopping = 'shopping',
    Social = 'social',
    Community = 'community',
    Volunteer = 'volunteer',
    Nature = 'nature',
    Fitness = 'fitness',
    Adventure = 'adventure',
    Relaxation = 'relaxation'
}